Remove redundant comments and split src.cpp to headers
