+++ src/src.vcxproj
@@ -140,12 +140,16 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="framework.h" />
+    <ClInclude Include="CmdArgumentParser.h" />
+    <ClInclude Include="FileInteractions.h" />
+    <ClInclude Include="HttpClientRequest.h" />
     <ClInclude Include="Resource.h" />
     <ClInclude Include="src.h" />
+    <ClInclude Include="StuffForWindow.h" />
     <ClInclude Include="targetver.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="FileInteractions.cpp" />
     <ClCompile Include="src.cpp" />
   </ItemGroup>
   <ItemGroup>
